Mapeo, tabla y nombre de columna de producto terminado

+++ TP2_Grupo21/src/main/java/ar/edu/unju/fi/service/IProductoService.java
@@ -2,7 +2,7 @@ package ar.edu.unju.fi.service;
 
 import java.util.List;
 
-import ar.edu.unju.fi.model.Producto;
+import ar.edu.unju.fi.entity.Producto;
 
 public interface IProductoService {
 	//Devuelve la lista de Productos
@@ -12,7 +12,7 @@ public interface IProductoService {
 	// Permite guardar los productos 
 	void guardar(Producto Producto);
 	//Permite realizar la busqueda mediante el codigo
-	Producto getBy(int codigo);
+	Producto getBy(Long codigo);
 	// Permite modificar los productos
 	void modificar (Producto productoModificado);
 	// Permite eliminar los productos
